style: #6 이메일 로그인 스타일 분리
