+++ src/components/emailLogin/IDInputbox.tsx
@@ -1,0 +1,17 @@
+import React, {Component} from 'react';
+import {TextInput, StyleSheet} from 'react-native';
+import styles from '../../styles/emailLogin/IDInputbox';
+
+class InputboxID extends Component {
+    render() {
+        return (
+            <TextInput
+                style = {styles.inputStyle}
+                placeholder = '아이디'
+            >
+            </TextInput>
+        );
+    }
+};
+
+export default InputboxID;
