+++ src/styles/emailLogin/IDInputbox.js
@@ -1,0 +1,13 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    inputStyle: {
+        marginRight: 20,
+        padding: 10,
+        backgroundColor: 'white',
+        borderColor: '#FFFFF',
+        borderRadius: 8
+    }
+});
+
+export default styles;
