+++ src/styles/emailLogin/LoginButton.js
@@ -1,0 +1,22 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    buttonStyle: {
+        marginTop: 20,
+        marginLeft: 20,
+        marginRight: 20,
+        padding: 10,
+        backgroundColor: '#2BC63B',
+        borderRadius: 8
+    },
+
+    buttonTitle: {
+        textAlign: 'center',
+        color: 'white', 
+        fontSize: 20,
+        fontFamily: 'GmarketSansTTFMedium'
+    }
+});
+
+
+export default styles;
