+++ src/styles/emailLogin/PWInputbox.js
@@ -1,0 +1,12 @@
+import {StyleSheet} from 'react-native';
+
+const styles = StyleSheet.create({
+    inputStyle: {
+        marginRight: 20,
+        padding: 10,
+        backgroundColor: 'white',
+        borderColor: '#FFFFF',
+        borderRadius: 10
+    }
+});
+export default styles;
