--- src/styles/emailLogin/SearchButton.js
@@ -1,10 +1,0 @@
-import {StyleSheet} from 'react-native';
-
-const styles = StyleSheet.create({
-    buttonStyle: {
-        marginLeft: 5,
-        padding: 10
-    }
-});
-
-export default styles;
